% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{as_df_res}
\alias{as_df_res}
\title{Observe deregulation between groups}
\usage{
as_df_res(res_dereg)
}
\arguments{
\item{res_dereg}{the output from \code{calc_dist}}
}
\value{
Results
}
\description{
Observe deregulation between groups
}
