% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_corr}
\alias{calc_corr}
\title{Compute the correlation between the gene expression in tumors and the micro-environment proportion}
\usage{
calc_corr(D_cancer, A)
}
\arguments{
\item{D_cancer}{The matrix with the gene expression in each tumor}

\item{A}{The matrix of the different cell type proportion in each sample}
}
\value{
A data.frame with for each gene and each cell type the correlation between the gene expression and the cell type proportion
}
\description{
Compute the correlation between the gene expression in tumors and the micro-environment proportion
}
\seealso{
\code{compute_1_res} \code{plot_res}
}
