% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_dist}
\alias{calc_dist}
\title{Compute distances between two groups: deregulated genes versus regulated genes}
\usage{
calc_dist(penda_res, matrix_A)
}
\arguments{
\item{penda_res}{Output from \code{pre_treat} function}

\item{matrix_A}{Matrix A of proportion of cell lines distribution per tumor \cr \cr  Dimensions: n(cell lines) x alpha(cell lines proportion per patient)}
}
\value{
Matrix of dimension geneNumber*cellLineNumber\cr \cr Each element is filled out by the four statistical tests: kanto, t-test, ks and cor
}
\description{
For each cell line and gene, four statistical tests are applied to evaluate the distance: \itemize{
\item{Kantorovich metric: Wasserstein distance, result : distance}
\item{Student test: mean comparison, result: -log10(p-value)}
\item{Kolmogorov-Smirnov: repartition comparison, result: distance and -log(p-value)}
\item{Correlation: Variability between two groups}}
}
