% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compute_1_res}
\alias{compute_1_res}
\title{P values and ROC curves}
\usage{
compute_1_res(values, genes, genes_fibro, genes_immune, pval)
}
\arguments{
\item{values}{statistical test used}

\item{genes}{gene names}

\item{genes_fibro}{gene expression in fibro cell line}

\item{genes_immune}{gene expression in immune cell line}

\item{pval}{list of p-values observed}
}
\value{
In a list: True positive, False positive, True negative and False negative with associated rates
}
\description{
P values and ROC curves
}
