% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_c}
\alias{corr_prop_c}
\title{Complex deregulations in tumors micro-environment}
\usage{
corr_prop_c(T_cancer, G, A, y = 2, thres_i = 0.1, thres_f = 0.45)
}
\arguments{
\item{T_cancer}{The matrix T, for more documentation : \code{simu_T_cancer}}

\item{G}{Gene number to be sampled, this function requires an \strong{even} Gene number}

\item{A}{The matrix A, for more documentation \code{simu_A}}

\item{y}{Overexpression coefficient to multiply the gene expression, {default} = 2}

\item{thres_i}{Threshold to induce overexpression in immune cells, {default} = 0.1}

\item{thres_f}{Threshold to induce overexpression in fibro cells, {default} = 0.45}
}
\value{
List with : \cr Deregulated matrix $T and deregulated genes in $g_immune and $g_fibro
}
\description{
Deregulations with a complex model in the matrix T generated from \code{simu_T_cancer} with cell lines proportions from the return of \code{simu_A} method \cr \cr
Step 1: G genes are drawn\cr \cr 
Step 2: Checking : proportions of cell type immune or fibro is respectively greater than \strong{tres_i} and \strong{tres_f} ? \cr \cr
Step 3: For the cell lines (fibro, immune) having a proportion greater than threshold: \cr \cr
newGeneExpression = geneExpression * y
}
