% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_f}
\alias{corr_prop_f}
\title{Simulation of the correlation between tumor micro-environment proportion and tumor gene expression based on a Factor}
\usage{
corr_prop_f(T_cancer, G, A_ME, z = 2)
}
\arguments{
\item{T_cancer}{The matrix of gene expression in each tumor, can be generated with: \code{simu_T_cancer}}

\item{G}{Number of genes correlated with the micro-environment composition}

\item{A_ME}{The vector with the proportion of the micro-environment in each tumor. Can be generated with a row of: \code{simu_A}}

\item{z}{Deregulation coefficient}
}
\value{
A list with : \cr Deregulated matrix $T and deregulated genes in $g_dereg
}
\description{
Generation of deregulations with a model based on factor in the matrix T generated from \code{simu_T_cancer} with cell lines proportions from the matrix A \code{simu_A} \cr \cr
Step 1: G genes are drawn\cr \cr 
Step 2: The over-expressions of G genes is induced by: \cr \cr
newGeneExpression = geneExpression x [1 + (z x cellTypeProportion)]
}
