% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_n}
\alias{corr_prop_n}
\title{Generator of new deregulations in tumors micro-environment}
\usage{
corr_prop_n(T_cancer, G, A, z = 2)
}
\arguments{
\item{T_cancer}{}

\item{G}{}

\item{A}{}

\item{z}{}
}
\value{

}
\description{
Generation of deregulations in the matrix T generated from @RiTMIC::simu_T_cancer() with cell lines proportions from the matrix A @RiTMIC::simu_A() \cr \cr
Step 1: G genes are drawn\cr \cr 
Step 2: For the cell lines, over-expressions to every G genes is induced by: \cr
newGeneExpression = geneExpression + [(1 + z)* cellTypeProportion)]
}
