% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_t}
\alias{corr_prop_t}
\title{Simulation of the correlation between gene expression and tumor 
micro-environment proportion based on a threshold.}
\usage{
corr_prop_t(T_cancer, A_ME, G = 100, dereg_coeff = 2, thres_ME = 0.1)
}
\arguments{
\item{T_cancer}{The matrix of gene expression in each tumor. Can be 
generated with: \code{simu_T_cancer}.}

\item{A_ME}{The vector with the proportion of the micro-environment in each tumor.
Can be generated with a row of: \code{simu_A}.}

\item{G}{The number of genes correlated with the micro-environment composition.}

\item{dereg_coeff}{The deregulation coefficient.}

\item{thres_ME}{The threshold to induce the overexpression.}
}
\value{
A list of $T the deregulated T matrix and $g_dereg the deregulated genes.
}
\description{
\code{corr_prop_t} simulates a link between the gene expression of \code{G} 
genes and the micro-environment (ME) proportion, only if this proportion is greater 
than \code{thres_ME}.
}
\details{
\code{G} genes are drawn, then for each tumor in \code{T_cancer}: \cr
If the proportion of the ME in this tumor is greater than \code{thres_ME}: \cr
New gene expression = gene expression * \code{dereg_coeff}.
}
