% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{eval_results}
\alias{eval_results}
\title{For one p-value, test relevance of the results.}
\usage{
eval_results(values, genes, genes_dereg, pval)
}
\arguments{
\item{values}{The vector of values associated at each gene, can be p-values 
of a statistical test, correlations, distances, etc. 
See also: \code{\link[Ritmic]{calc_dist}}}

\item{genes}{The vector with the names of the genes.}

\item{genes_dereg}{The vector with the names of the genes deregulated in the
simulations.}

\item{pval}{The threshold of interest, adapted for the metric used in values.}
}
\value{
A vector with the number of True Positive, False Positive, True 
  Negative and False Negative results, the False Positive Rate and the True 
  Positive Rate.
}
\description{
\code{eval_results} compares the computed results and the genes deregulated 
in the simulation to obtain the confusion matrix (TP, FP, TN, FN) necessary to 
plot ROC curves.
}
