% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_D}
\alias{matrix_D}
\title{Expression matrix of cases processed with simu_D with noise=T}
\format{
A matrix with 3000 rows and 46 columns
}
\source{
\code{simu_D}
}
\usage{
matrix_D
}
\description{
A matrix...
}
\keyword{datasets}
