% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_D2}
\alias{matrix_D2}
\title{Expression matrix of cases with noise=F}
\format{
A matrix with 3000 rows and 46 columns
}
\source{
\code{simu_D}
}
\usage{
matrix_D2
}
\description{
A matrix...
}
\keyword{datasets}
