% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pre_plot_res}
\alias{pre_plot_res}
\title{Load environment required to \code{plot_res} function}
\usage{
pre_plot_res(matrix_T, matrix_A, compute_1_res_output)
}
\arguments{
\item{matrix_T}{The matrix_T output from corr_prop functions}

\item{matrix_A}{The matrix_A from \code{simu_A}}

\item{compute_1_res_output}{output from \code{compute_1_res} function}
}
\value{
correlation values between gene expressions
}
\description{
Load environment required to \code{plot_res} function
}
\seealso{
\code{plot_res}
}
