% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pre_treat}
\alias{pre_treat}
\title{Pre-treatment for the link between gene deregulation and microenvironment composition}
\usage{
pre_treat(penda_res, thres_p)
}
\arguments{
\item{penda_res}{The result of the Penda method with the deregulation of each gene for each sample  \code{\link[penda]{penda_test}}}

\item{thres_p}{Minimum number of samples for the deregulated and not deregulated groups.}
}
\value{
A binary matrix with the genes that can be used for the statistical analysis and their deregulation status for each sample.
}
\description{
Matrix are sorted to conserved only the genes with a different deregulation statut in more than thres_p samples
}
