% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treat_Dcancer.R
\name{pre_treat_Dcancer}
\alias{pre_treat_Dcancer}
\title{Pretreatment of the Dcancer matrix}
\usage{
pre_treat_Dcancer(Dcancer, A_cancer, prop_cancer_threshold = 0.1)
}
\arguments{
\item{Dcancer}{The matrix of gene expression in each cancer sample.}

\item{A_cancer}{A vector with the cancer cell proportion in each sample.}

\item{prop_cancer_threshold}{The minimum proportion of cancer cells needed to keep the sample.}
}
\value{
The Dcancer matrix without negative expression and samples with a 
cancer proportion under \code{prop_cancer_threshold}.
}
\description{
\code{pre_treat_Dcancer} pretreat the Dcancer matrix to remove genes with
a negative expression and samples with a proportion of cancer cells under
\code{prop_cancer_threshold}.
}
