% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_D}
\alias{simu_D}
\title{Build the D matrix: gene expression per sample}
\usage{
simu_D(A, T, T_cancer, noise = F, mean = 0, sd = 0.1, val_min = 0, val_max = 1)
}
\arguments{
\item{A}{The matrix of the different cell type proportion in each sample, can be generated with: \code{simu_A}. Note: the last row will be used for the tumor.}

\item{T}{The matrix with the gene expression profile of each cell type, except tumors}

\item{T_cancer}{The matrix of gene expression in each tumor, can be generated with: \code{simu_T_cancer}}

\item{noise}{TRUE for adding a Gaussian noise on the product matrix, FALSE for not (see \code{add_noise})}

\item{mean}{Mean of the noise (see \code{add_noise})}

\item{sd}{Standard deviation of the noise (see \code{add_noise})}

\item{val_min}{Minimum value of the matrix after the noise (see \code{add_noise})}

\item{val_max}{Maximum value of the matrix after the noise (see \code{add_noise})}
}
\value{
The D matrix of size n*p with the complex gene expression for each sample. If noise = T, return a list of the D matrix before the noise (D) and after the noise (D_noise)
}
\description{
Simulation of a complex RNAseq matrix, obtained with the matrix multiplication of the A and T matrices, with a different tumor profile for each sample
}
