% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_noise}
\alias{add_noise}
\title{Add a Gaussian noise to a matrix.}
\usage{
add_noise(D, mean = 0, sd = 0.1, val_min = 0, val_max = 1)
}
\arguments{
\item{D}{The matrix where to add a noise.}

\item{mean}{The mean of the noise.}

\item{sd}{The standard deviation of the noise.}

\item{val_min}{The minimum value for the final matrix.}

\item{val_max}{The maximum value of the final matrix.}
}
\value{
A matrix corresponding to the D matrix with a gaussian noise.
}
\description{
\code{add_noise} Apply a gaussian noise with a lower cap and an upper cap for 
the values.
}
