% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to the simulations of tumor micro-environments}
\usage{
add_noise(data, mean = 0, sd = 0.1, val_min = 0, val_max = 1)
}
\arguments{
\item{data}{matrix_T with or without deregulations}

\item{mean}{}

\item{sd}{}

\item{val_min}{}

\item{val_max}{}
}
\value{

}
\description{
By adding noise, the simulation will have a better reflection of the biological simulation
}
\seealso{
RiTMIC::simu_T_cancer()

RiTMIC::corr_prop functions...
}
