% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to the analysis or simulations of the tumor micro-environments gene expressions}
\usage{
add_noise(data, mean = 0, sd = 0.1, val_min = 0, val_max = 1)
}
\arguments{
\item{data}{matrix_D with or without deregulations obtained from the combinations of matrix_T * matrix_A}

\item{mean}{}

\item{sd}{}

\item{val_min}{}

\item{val_max}{}
}
\value{

}
\description{
Applications of adding noise : \cr Simulations : Better reflection of the biological observations \cr
Biological observation : improve the generalization error to avoid over-learning
}
\seealso{
RiTMIC::simu_T_cancer

RiTMIC::corr_prop functions...
}
