% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{as_df_res}
\alias{as_df_res}
\title{Find cell marker from deregulated genes between cell lines}
\usage{
as_df_res(res_dereg)
}
\arguments{
\item{res_dereg}{Output from \code{calc_dist} function}
}
\value{
Data_frame with accurate
}
\description{
Find specific cell line markers by observe a high deregulation between a single cell line and the others
}
\seealso{
\code{calc_dist}
}
