% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_corr}
\alias{calc_corr}
\title{Load environment required to \code{plot_res} function}
\usage{
calc_corr(matrix_T_control, matrix_A)
}
\arguments{
\item{matrix_T_control}{The matrix_T output from corr_prop functions}

\item{matrix_A}{The matrix_A from \code{simu_A}}
}
\value{
correlation values between gene expressions
}
\description{
Load environment required to \code{plot_res} function
}
\seealso{
\code{compute_1_res} \code{plot_res}
}
