% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_dist}
\alias{calc_dist}
\title{Compute metrics between the two groups: samples with deregulated genes versus not deregulated genes}
\usage{
calc_dist(binary_penda, A)
}
\arguments{
\item{binary_penda}{The binary matrix of deregulation, can be obtain with the \code{pre_treat} function}

\item{A}{The matrix of the different cell type proportion in each sample \cr \cr  Dimensions: k (cell types) * p (samples)}
}
\value{
A data.frame with for each gene (genes) and each celle type (type), the result of the 3 metrics
}
\description{
For each cell type and each gene, 3 statistical tests are applied to evaluate the distances bewteen the cell type proportion of the two groups of deregulation: 
\tabular{lll}{
Statistical test   \tab Tested parameter               \tab Result type\cr 
Kantorovich        \tab Kantorovich distance           \tab distance\cr
Student t-test     \tab Mean comparison                \tab -log10(p-value)\cr
Kolmogorov-Smirnov \tab Repartition comparison         \tab -log(p-value)\cr
}
}
