% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_dist}
\alias{calc_dist}
\title{Compute distances between two groups: deregulated genes versus regulated genes}
\usage{
calc_dist(penda_res, matrix_A)
}
\arguments{
\item{penda_res}{output from \code{pre_treat} function}
}
\value{
Matrix of dimension geneNumber*cellLineNumber\cr Each element is filled by the four statistical tests: kanto, t-test, ks and cor
}
\description{
For each cell line and gene, four statistical tests are applied to evaluate the distance: \itemize{
\item{Kantorovich metric: Wasserstein distance, result : distance}
\item{Student test: mean comparison, result: -log10(p-value)}
\item{Kolmogorov-Smirnov: repartition comparison, result: distance and -log(p-value)}
\item{Correlation: Variability between two groups}}
}
