% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_dist}
\alias{calc_dist}
\title{Comparison between gene deregulation and cell type proportion.}
\usage{
calc_dist(binary_penda, A)
}
\arguments{
\item{binary_penda}{The binary matrix of deregulation 0/1 for each gene and 
each sample, can be obtain with the \code{pre_treat} function. 
Dimension n (genes) * p (samples).}

\item{A}{The matrix of cell type proportions for each sample. 
Dimension p (samples) * k (cell types).}
}
\value{
A data frame with for each gene and each cell type the results of the 3 metrics.
}
\description{
\code{calc_dist} groups samples according to their deregulation status 
for a given gene. It then computes for each cell type different metrics to 
evaluate the distance between the two groups of cell proportions. An important 
distance suggests a link between the gene expression and the cell type proportion. \cr 
Metrics: 
\tabular{lll}{
\strong{Metric}          \tab \strong{Tested parameter} \tab \strong{Computed parameter}\cr 
Kantorovich distance     \tab Transportation           \tab Distance\cr
Student's t-test         \tab Mean comparison          \tab -log10(p-value)\cr
Kolmogorov-Smirnov test  \tab Distribution comparison  \tab -log10(p-value)\cr
}
}
