% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_dist}
\alias{calc_dist}
\title{Compute distances between two groups: deregulated genes versus regulated genes}
\usage{
calc_dist(penda_res, matrix_A)
}
\arguments{
\item{penda_res}{Output from \code{pre_treat} function}

\item{matrix_A}{Matrix A of proportion of cell lines distribution per tumor \cr \cr  Dimensions: n(cell lines) x alpha(cell lines proportion per patient)}
}
\value{
Matrix of dimension geneNumber*cellLineNumber\cr \cr Each element is filled out by the four statistical tests: kanto, t-test, ks and cor
}
\description{
For each cell line and gene, four statistical tests are applied to evaluate the distance: 
\tabular{lll}{
Statistical test   \tab Tested parameter               \tab Result type\cr 
Kantorovich metric \tab Wasserstein distance           \tab distance\cr
Student test       \tab Mean comparison                \tab -log10(p-value)\cr
Kolmogorov-Smirnov \tab Repartition comparison         \tab distance and   -log(p-value)\cr
Correlation        \tab Variability between two groups \tab Correlation coefficient 
}
}
