% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_c}
\alias{corr_prop_c}
\title{Complex deregulations in tumors micro-environment}
\usage{
corr_prop_c(T_cancer, G, A, y = 2, thres_i = 0.1, thres_f = 0.45)
}
\arguments{
\item{T_cancer}{The matrix T, for more documentation @RiTMIC::simu_T_cancer()}

\item{G}{Gene numbers to be drawn}

\item{A}{The matrix A, for more documentation @RiTMIC::simu_A()}

\item{y}{Overexpression coefficient to multiply the gene expression, by default y = 2}

\item{thres_i}{Threshold to induce overexpression in immune cells, by default thres_i = 0.1}

\item{thres_f}{Threshold to induce overexpression in fibro cells, by default thres_f = 0.45}
}
\value{
A list with : Deregulated matrix $T and deregulated genes in $g_immune and $g_fibro
}
\description{
Generation of deregulations with a complex model in the matrix T generated from @RiTMIC::simu_T_cancer() with cell lines proportions from the matrix A @RiTMIC::simu_A() \cr \cr
Step 1: G genes are drawn\cr \cr 
Step 2: Checking : proportions of cell type immune or fibro is respectively greater than @tres_i and @tres_f ? \cr \cr
Step 3: For the cell lines (fibro, immune) having a proportion greater than threshold: \cr \cr
newGeneExpression = geneExpression * y
}
