% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_n}
\alias{corr_prop_n}
\title{New deregulations in tumors micro-environment}
\usage{
corr_prop_n(T_cancer, G, A, z = 2)
}
\arguments{
\item{T_cancer}{The matrix T, for more documentation \code{simu_T_cancer}}

\item{G}{Gene number to be sampled, this function requires an \strong{even} Gene number}

\item{A}{The matrix A, for more documentation \code{simu_A}}

\item{z}{Overexpression coefficient to multiply the gene expression, by default z = 2}
}
\value{
A list with : \cr Deregulated matrix $T and deregulated genes in $g_immune and $g_fibro
}
\description{
Generation of deregulations from a new model in the matrix T generated from \code{simu_T_cancer} with cell lines proportions from the matrix A \code{simu_A} \cr \cr
Step 1: G genes are drawn\cr \cr 
Step 2: For the cell lines, over-expressions to every G genes is induced by: \cr \cr
newGeneExpression = geneExpression + [(1 + z)* cellTypeProportion)]
}
