% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{corr_prop_s}
\alias{corr_prop_s}
\title{Simple deregulations in tumors micro-environment}
\usage{
corr_prop_s(T_cancer, G, A, x = 100)
}
\arguments{
\item{T_cancer}{RNAseq data}

\item{G}{Gene number to be sampled, this function requires an \strong{even} Gene number}

\item{A}{Matrix A : distribution of cell lines per tumor \code{simu_A}}

\item{x}{Coefficient, by default it's settled to 100}
}
\value{
List with : \cr Deregulated matrix $T and deregulated genes in $g_immune and $g_fibro
}
\description{
Generation of deregulations with a simple model in the matrix T generated from \code{simu_T_cancer} with cell lines proportions from the return of \code{simu_A} method \cr\cr
Step 1: G genes are drawn\cr \cr 
Step 2: Over-expression induced to these genes : \cr \cr
newGeneExpression = geneExpression + x * cellTypeProportion
}
