% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{eval_results}
\alias{eval_results}
\title{For one p-value, test if the we find genes deregulated in the simulations}
\usage{
eval_results(values, genes, genes_dereg, pval)
}
\arguments{
\item{values}{the result of the test for each gene (ex: distance value, p-value, etc.)}

\item{genes}{names of the genes analysed}

\item{genes_dereg}{names of the genes deregulated in the simulations}

\item{pval}{the threshold of interest, can be a -log10(pval) or a distance}
}
\value{
A vector with the number of TP, FP, TN, FN, the FPR and the TPR
}
\description{
Compute the confusion matrix(TP,FP,TN,FN) necessary to plot ROC curves
}
