% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_data_bypatient}
\alias{generate_data_bypatient}
\title{Compute the down and up regulated genes in the two lists H and L}
\usage{
generate_data_bypatient(D_list, U_list)
}
\arguments{
\item{D_list}{}

\item{U_list}{}
}
\value{

}
\description{
Compute the down and up regulated genes in the two lists H and L
}
