% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_res}
\alias{plot_res}
\title{Plot the ROC curve of the detection of deregulated genes linked to micro-environment.}
\usage{
plot_res(
  corr_res,
  dist_res,
  genes_dereg,
  pvalues = c(0, 0.01, 0.05, 0.1, 1),
  graph_title = ""
)
}
\arguments{
\item{corr_res}{The data frame output from \code{\link[Ritmic]{calc_corr}},
with the correlation for each gene of the cell type deregulated.}

\item{dist_res}{The data frame output from \code{\link[Ritmic]{calc_dist}}, 
with the metrics for each gene of the cell type deregulated.}

\item{genes_dereg}{The vector with the names of genes deregulated in the simulation.}

\item{pvalues}{The vector of thresholds to trace the ROC curve.}

\item{graph_title}{An optional title for the ROC curve plot.}
}
\value{
A ggplot object with the ROC curves. See \code{\link[ggplot2]{ggplot}}
}
\description{
\code{plot_res} makes a ROC plot with the detection of genes deregulated in 
simulations. It makes one curve for the direct correlation between genes 
expression and micro-environment proportion, and three curves for the comparison
between PenDA results and micro-environment, with three metrics: \cr 
\itemize{
  \item{The pvalue of the Kolmogorov-Smirnov test} \item{The p-value of the Student's t-test} \item{The distance of Kantorovich}
}
}
\seealso{
\code{\link[Ritmic]{compute_1_res}}
}
