% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_res}
\alias{plot_res}
\title{Check the PenDA enrichment: Plot ROC curves}
\usage{
plot_res(
  corr_res,
  dist_res,
  genes_dereg,
  pvalues = c(0, 0.1, 0.5, 0.8),
  graph_title
)
}
\arguments{
\item{corr_res}{Output from \code{calc_corr} function for the cell type deregulated}

\item{dist_res}{Output from \code{calc_dist} function for the cell type deregulated}

\item{genes_dereg}{The names of genes deregulated in the simulation}

\item{pvalues}{The threshold for the ROC curve}

\item{graph_title}{Title of the ROC curve graph}
}
\value{
ROC curves
}
\description{
Compare non-enriched data by PenDA with a correlation test versus the PenDA efficiency with tests: \cr \itemize{
\item{Kolmogorov-Smirnov} \item{Student} \item{Kantorovich}
}
}
\seealso{
\code{as_def_res} \code{pre_plot_res} \code{compute_1_res}
}
