% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{plot_res}
\alias{plot_res}
\title{Check the PenDA enrichment: Plot ROC curves}
\usage{
plot_res(pre_plot_res_output, calc_dist_output, graph_title)
}
\arguments{
\item{pre_plot_res_output}{Output from \code{pre_plot_res} function}

\item{graph_title}{Title of the ROC curve graph}
}
\value{
ROC curves
}
\description{
Compare non-enriched data by PenDA with a correlation test versus the PenDA efficiency with tests: \cr \itemize{
\item{Kolmogorov-Smirnov} \item{Student} \item{Kantorovich}
}
}
\seealso{
\code{as_def_res} \code{pre_plot_res}
}
