% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{plot_res}
\alias{plot_res}
\title{Plot ROC curves to see PenDA improvement}
\usage{
plot_res(T_matrix, df_res, pre_plot_res, graph_title)
}
\arguments{
\item{T_matrix}{Matrix T output from corr_prop functions}

\item{df_res}{Output from \code{as_df_res} function}

\item{pre_plot_res}{Output from \code{pre_plot_res} function}

\item{graph_title}{The title of the ROC curve graph}
}
\value{
ROC curves
}
\description{
Compare the panda efficiency with tests: Kolmogorov-Smirnov, Student and Kantorovich versus a correlation test
}
\seealso{
\code{as_def_res} \code{pre_plot_res}
}
