% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pre_treat}
\alias{pre_treat}
\title{Pre-treatment for statistical analysis}
\usage{
pre_treat(penda_res, patientNumber)
}
\arguments{
\item{penda_res}{Output from \code{\link[penda]{penda_test_1ctrl}}}

\item{patientNumber}{Minimum number of patients with the gene deregulation}
}
\value{
Matrix without non deregulated genes
}
\description{
Matrix are sorted to have the common patients with only genes that are deregulated in more than a minimum number of patients
}
