% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_A}
\alias{simu_A}
\title{Simulation of the A matrix}
\usage{
simu_A(n, alpha = c(1.5, 4.5, 1, 3))
}
\arguments{
\item{n}{The number of samples.}

\item{alpha}{The vector of size kwith the different cell types proportion.}
}
\value{
This function return a matrix of size k*p with the cell type proportions.
}
\description{
\code{simu_A} generates the \strong{A matrix}: \cr 
Different cell types proportions in each sample. \cr 
Dimensions: k*p
}
\details{
The cell type distribution in a sample is simulated by a Dirichlet 
distribution. See: \code{\link[gtools]{rdirichlet}}
}
\seealso{
\code{\link[gtools]{rdirichlet}}
}
