% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_A}
\alias{simu_A}
\title{Simulate cell lines proportion distribution in tumors}
\usage{
simu_A(
  n,
  alpha = c(1.5, 4.5, 1, 3),
  nameCellLines = c("normal", "fibro", "immune", "tumor")
)
}
\arguments{
\item{n}{patients number}

\item{alpha}{cell type proportions of size k in a concatenated vector}
}
\value{
Proportion of different cell lines in a matrix
}
\description{
\code{simu_A} generates the \strong{matrix A} : \cr Proportion of cell lines distribution per tumor \cr \cr  Dimensions: n x alpha
}
\details{
Cell lines distributions are obtained with \code{\link[gtools]{rdirichlet}} method
}
\examples{
simu_A(20,c(1.5, 4.5, 3))
}
\seealso{
\code{\link[gtools]{rdirichlet}}
}
