% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_A}
\alias{simu_A}
\title{Simulate cell line distributions in tumors}
\usage{
simu_A(n, alpha = c(1.5, 4.5, 1, 3))
}
\arguments{
\item{n}{patients number}

\item{alpha}{cell type proportions of size k in a concatenated vector}
}
\value{
Proportion of different cell lines in a matrix
}
\description{
Creates a new matrix with @n : the patient number and @alpha : the vector of cell lines proportion of size k
}
\examples{
simu_A(20,c(1.5, 4.5, 3))
}
