% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_A}
\alias{simu_A}
\title{Title simu_A}
\usage{
simu_A(n, alpha = c(1.5, 4.5, 1, 3))
}
\arguments{
\item{n}{number of patients}

\item{alpha}{proportions of different cell types}
}
\value{
the matrix of proportions
}
\description{
Title simu_A
}
\examples{
simu_A(n = 30, alpha = c(1.5, 4.5, 1, 3))
}
