% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_D}
\alias{simu_D}
\title{Simulation of an RNAseq matrix, obtained with the matrices combinations of the matrix A and T for each sample}
\usage{
simu_D(
  matrix_A,
  matrix_T,
  corr_matrix_T,
  noise = T,
  mean = 0,
  sd = 0.1,
  val_min = 0,
  val_max = 1
)
}
\arguments{
\item{matrix_A}{matrix obtained with \code{simu_A}}

\item{matrix_T}{matrix obtained with \code{simu_T_cancer}}

\item{noise}{Boolean statement, {default} = T \cr\cr Do you want to use \code{add_noise} to your resulting matrix to increase the variance ?}

\item{mean}{Mean of the noise, {default} = 0}

\item{sd}{Standard deviation of the noise, {default} = 0.1}

\item{val_min}{Minimum value of the noise, {default} = 0}

\item{val_max}{Maximum value of the noise, {default} = 1}
}
\value{
matrix D of gene expression per sample
}
\description{
Simulation of an RNAseq matrix, obtained with the matrices combinations of the matrix A and T for each sample
}
\seealso{
\code{\link[RiTMIC]{add_noise}}
}
