% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_T_cancer}
\alias{simu_T_cancer}
\title{Simulate a tumor micro-environment}
\usage{
simu_T_cancer(tumor_RNAseq, n)
}
\arguments{
\item{tumor_RNAseq}{RNAseq dataset to simulate a random T matrix}

\item{n}{Supposed cell types number}
}
\value{
A matrix : Gene expression distribution per cell type
}
\description{
Generate a matrix with random gene expression distributions in n tumor cell lines with genes imported from RNA_seq dataset
}
