% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_T_cancer}
\alias{simu_T_cancer}
\title{Simulate a tumor micro-environment}
\usage{
simu_T_cancer(tumor_RNAseq, n)
}
\arguments{
\item{tumor_RNAseq}{RNAseq dataset to simulate a random T matrix}

\item{n}{Supposed cell types number}
}
\value{
Matrix A: Gene expressions distributions per cell line
}
\description{
Generate a \strong{matrix T} with random \strong{gene expression} distributions in \strong{n} tumor cell lines \cr Matrix dimension : n x gene_name
}
\details{
gene_name is imported from the tumor_RNAseq data \cr Random gene expression distributions is performed by \code{\link{rnorm}} \cr Probability of gene expression in a cell line is computed by \code{\link[mixtools]{normalmixEM}}
}
